Security: close identity-override + CSRF gaps in the Studio-era surface
Security review of everything added since the last audit (workspaces, team
invites, branded email, notifications). Findings and fixes:

HIGH — sendQuote server action accepted an identity override.
  `sendQuote(quoteId, { serviceContext: { userId } })` was an exported
  "use server" action, so the userId parameter was network-reachable: a
  caller could supply any userId and the action would run against the
  RLS-bypassing service client, sending that user's draft quotes to their
  clients (and flipping quote status). Split the logic into
  lib/quotes/send-quote-core.ts (not an action — no "use server", so the
  identity param is unreachable from the network). The action now derives
  identity from the session only; the workflow engine calls the core
  directly with the service client and the job row's user_id.

MEDIUM — send-invite trusted a body-supplied inviterEmail, letting any
  authenticated user send release invites appearing to come from anyone.
  Now derived from the session user. Callers updated.

MEDIUM — admin/log-activity had no replay guard on "signup". Any signed-in
  user could re-post {eventType:"signup"} repeatedly to re-send themselves
  welcome mail and spam every admin. Now only fires within 1h of account
  creation and only if no welcome email was ever logged for that user.

MEDIUM — cron/check-distribution didn't fail closed when CRON_SECRET is
  unset (the expected digest became "Bearer undefined", which is
  guessable). Matches the guard weekly-digest already had.

MEDIUM — 17 cookie-authenticated mutating routes were missing
  requireSameOrigin, the project's own CSRF convention (CLAUDE.md).
  Added to attribute-signup, calendar/generate-token, convert,
  integrations connect/disconnect, notify, perf/ingest, the 4 distribution
  routes, send-invite, and the 5 stripe checkout/portal/connect routes.
  Deliberately NOT added to token- or signature-authenticated endpoints
  (portal/*, webhook, unsubscribe, OAuth callback) — they have no Origin.

LOW — portal/comment DELETE compared delete_token with !== (timing
  oracle). Now sha256 + timingSafeEqual, same as the cron secret fix.

LOW — csp-report had no rate limit despite its docblock claiming one;
  it's unauthenticated and logs attacker-controlled JSON. Now 20/min/IP.

Migration 080 — align EXECUTE revokes on SECURITY DEFINER functions.
  Later migrations relied on CREATE OR REPLACE preserving grants and each
  omitted a leg; anon still had EXECUTE on accessible_release_ids,
  admin_workspace_ids, handle_new_user, stamp_owner_workspace_id and
  sync_workspace_plan_from_subscription. APPLIED TO PROD and verified:
  anon now false on all 9; authenticated retains EXECUTE on exactly the 5
  RLS helpers that need it (release access + invite claim still resolve).

No permission changes for legitimate users: every fix either removes an
override that only an attacker could use, or adds a check the app's own
same-origin fetches already satisfy.

tsc clean; eslint clean apart from the pre-existing require("resend")
lazy-init pattern.
