+++ src/app/api/releases/[releaseId]/distribution/[distId]/route.ts
@@ -1,12 +1,16 @@
 import { NextRequest, NextResponse } from "next/server";
 import { createSupabaseServerClient } from "@/lib/supabaseServerClient";
 import { rateLimit, getClientIp } from "@/lib/rate-limit";
+import { requireSameOrigin } from "@/lib/origin-check";
 
 type RouteContext = {
   params: Promise<{ releaseId: string; distId: string }>;
 };
 
 export async function PATCH(request: NextRequest, { params }: RouteContext) {
+  const originErr = requireSameOrigin(request);
+  if (originErr) return originErr;
+
   const ip = getClientIp(request);
   const { success } = rateLimit(`dist-update:${ip}`, 20, 60_000);
   if (!success) {
@@ -72,6 +76,9 @@ export async function PATCH(request: NextRequest, { params }: RouteContext) {
 }
 
 export async function DELETE(request: NextRequest, { params }: RouteContext) {
+  const originErr = requireSameOrigin(request);
+  if (originErr) return originErr;
+
   const ip = getClientIp(request);
   const { success } = rateLimit(`dist-delete:${ip}`, 20, 60_000);
   if (!success) {
