+++ src/app/api/releases/[releaseId]/distribution/bulk-submit/route.ts
@@ -2,12 +2,16 @@ import { NextRequest, NextResponse } from "next/server";
 import { createSupabaseServerClient } from "@/lib/supabaseServerClient";
 import { rateLimit, getClientIp } from "@/lib/rate-limit";
 import { PLATFORMS } from "@/lib/distribution/platforms";
+import { requireSameOrigin } from "@/lib/origin-check";
 
 type RouteContext = { params: Promise<{ releaseId: string }> };
 
 const validPlatforms = new Set<string>(PLATFORMS.map((p) => p.id));
 
 export async function POST(request: NextRequest, { params }: RouteContext) {
+  const originErr = requireSameOrigin(request);
+  if (originErr) return originErr;
+
   const ip = getClientIp(request);
   const { success } = rateLimit(`dist-bulk:${ip}`, 10, 60_000);
   if (!success) {
