+++ src/app/api/stripe/checkout/route.ts
@@ -3,6 +3,7 @@ import { createSupabaseServerClient } from "@/lib/supabaseServerClient";
 import { stripe } from "@/lib/stripe-server";
 import { getStripePriceId, type BillingInterval, type PaidPlan } from "@/lib/pricing";
 import { rateLimit, getClientIp } from "@/lib/rate-limit";
+import { requireSameOrigin } from "@/lib/origin-check";
 
 /**
  * POST /api/stripe/checkout
@@ -12,6 +13,9 @@ import { rateLimit, getClientIp } from "@/lib/rate-limit";
  *   - interval: "monthly" | "annual" (default: "monthly")
  */
 export async function POST(req: NextRequest) {
+  const originErr = requireSameOrigin(req);
+  if (originErr) return originErr;
+
   const ip = getClientIp(req);
   const { success } = rateLimit(`stripe-checkout:${ip}`, 10, 60_000);
   if (!success) return NextResponse.json({ error: "Too many requests" }, { status: 429 });
